backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.29.0
SQLAlchemy>=2.0
asyncpg
alembic
PyJWT
passlib
pydantic==1.10.13
pymongo
python-dotenv
python-multipart
redis
celery[redis]==5.3.6
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asyncpg
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-celery
opentelemetry-instrumentation-logging
opentelemetry-instrumentation-sqlalchemy
aiobreaker
